requirements.txt
flask==3.0.0
gunicorn==21.2.0
openai==1.54.0
playwright==1.48.0
requests==2.32.0
beautifulsoup4==4.12.3
pandas==2.2.3
numpy==2.1.3
pillow==11.0.0
PyPDF2==3.0.1
python-dotenv==1.0.1
lxml==5.3.0
openpyxl==3.1.5
matplotlib==3.9.2
seaborn==0.13.2
aiohttp==3.10.10
pydantic==2.10.0
